+++ ProjektODPP/datareader.h
@@ -10,7 +10,7 @@ public:
                       int Czas1, int Czas2, int Czas3, int Czas4);
     bool Odczyt_Pliku(std::string nazwa_pliku);
     bool Formatowanie(std::string nazwa_pliku, std::vector <unsigned int> Permutacje);
-    bool Wyniki(std::string nazwa_pliku, std::vector <int>  Start, std::vector <int> Koniec, std::vector <unsigned int> Permutacje);
+    bool Wyniki(std::string nazwa_pliku, std::vector <double>  Start, std::vector <double> Koniec, std::vector <unsigned int> Permutacje);
     bool Stworz_Wektor_In(std::string nazwa_pliku);
     bool Stworz_Wektor_Out(std::string nazwa_pliku);
     DataReader();
